making cart class iterable and showing the details

+++ Ecommerce/home/models.py
@@ -27,7 +27,7 @@ class Product(models.Model):
     slug = models.SlugField(max_length=255, unique=True)
     image = models.ImageField()
     description = models.TextField()
-    price = models.DecimalField(max_digits=100, decimal_places=2)
+    price = models.IntegerField()
     available = models.BooleanField(default=True)
     created = models.DateTimeField(auto_now_add=True)
     updated = models.DateTimeField(auto_now=True)
